Add an action to label draft PRs
This took a bit of back-and-forth, since most people seem to use
templates or scripts for this.

Fortunately, though I learned a bit...
- About the github CLI
- About permissions for actions
- About errors actions can throw

I think I should be able to add a ready_for_review job to this same
action to remove the "WIP" and put on "Ready for Review".

Then, I wonder if I should look into scripts w/ TS for adding/removing
labels?